Regexp groups less eager: support multiple links
Previous regular expression detecting links was too eager capturing the contents inside the parens (`()`)  causing the `replaceAllIn` to effectively replace only the last occurrence in a String. With this regexp all links in a string are captured.

With that change we don't need to split the body `s` into lines just to recompose it back.